% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create.info.mask}
\alias{create.info.mask}
\title{Determine which time periods have information from the CMS sensors and which do not}
\usage{
create.info.mask(times, sims, gap.time = 0, length.threshold = 15)
}
\description{
Determine which time periods have information from the CMS sensors and which do not
}
