% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_durations.R
\name{estimate_durations}
\alias{estimate_durations}
\title{Estimate emission durations using CMS, accounting for CMS non-detect times}
\usage{
estimate_durations(data, directory, events, plot, time)
}
\arguments{
\item{data}{List. Forward model output data. Format must match output from MAIN_1_simulate.R, which can be found at: \url{https://github.com/wsdaniels/DLQ/}. Default set to shortened "toy" set from the 2023 ADED controlled release experiment (embedded in \code{duration} package).}

\item{directory}{Character. Location where results output is saved. If not specified, a directory is created and the location is shared with the user.}

\item{events}{Logical. Print the number of events per year? Default set to \code{FALSE}.}

\item{plot}{Logical. Return a plot of the distribution of average duration estimates by source? Default set to \code{FALSE}.}

\item{time}{Logical. Return the timing of the simulation. Default set to \code{TRUE}.}
}
\value{
Listed output estimated durations by source. Saved in a newly created directory (\code{Results Output}).
}
\description{
Estimate emission durations using CMS, accounting for CMS non-detect times
}
\examples{
estimate_durations(directory = "Desktop", plot = TRUE)
}
\author{
William Daniels, Philip Waggoner, Dorit Hammerling
}
