% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find.spikes}
\alias{find.spikes}
\title{Find spikes from DLQ code}
\usage{
find.spikes(
  times,
  obs,
  going.up.threshold = 0.25,
  return.threshold = 5,
  amp.threshold = 1,
  cont.diff.threshold = 0.25,
  cont.diff.num = 10
)
}
\description{
Find spikes from DLQ code
}
