% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.combine.info}
\alias{get.combine.info}
\title{Determine which naive events have non-zero probability of being combined with their neighbors}
\usage{
get.combine.info(spikes, info.list, loc.est.all.events, tz)
}
\description{
Determine which naive events have non-zero probability of being combined with their neighbors
}
