% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.durations}
\alias{get.durations}
\title{Get distribution of possible durations for a given naive event}
\usage{
get.durations(spikes, info.list, loc.est.all.events, rate.est.all.events, tz)
}
\description{
Get distribution of possible durations for a given naive event
}
