% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.event.counts}
\alias{get.event.counts}
\title{Get number of events by source, accounting for possibility to recombine events}
\usage{
get.event.counts(
  spikes,
  info.list,
  loc.est.all.events,
  rate.est.all.events,
  tz
)
}
\description{
Get number of events by source, accounting for possibility to recombine events
}
