% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perform.event.detection.alt}
\alias{perform.event.detection.alt}
\title{Create naive emission durations that do not account for CMS non-detect times
(Alt version for direct handling of first obs for testing with toy data)}
\usage{
perform.event.detection.alt(times, max.obs, gap.time, length.threshold)
}
\description{
Create naive emission durations that do not account for CMS non-detect times
(Alt version for direct handling of first obs for testing with toy data)
}
