% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perform.event.detection}
\alias{perform.event.detection}
\title{Create naive emission durations that do not account for CMS non-detect times}
\usage{
perform.event.detection(times, max.obs, gap.time, length.threshold)
}
\description{
Create naive emission durations that do not account for CMS non-detect times
}
