% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perform.localization}
\alias{perform.localization}
\title{Estimate emission source for each naive event}
\usage{
perform.localization(spikes, obs, sims)
}
\description{
Estimate emission source for each naive event
}
