% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perform.quantification}
\alias{perform.quantification}
\title{Estimate emission rate for each naive event}
\usage{
perform.quantification(times, spikes, obs, sims, loc.est.all.events)
}
\description{
Estimate emission rate for each naive event
}
