% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_durations}
\alias{plot_durations}
\title{Plot distributions of avg durations by source}
\usage{
plot_durations(durations_list, type = c("histogram", "density"))
}
\description{
Plot distributions of avg durations by source
}
