% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove.background}
\alias{remove.background}
\title{Remove background from concentration time series}
\usage{
remove.background(times, obs, going.up.threshold, amp.threshold, gap.time)
}
\description{
Remove background from concentration time series
}
