% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale.sims}
\alias{scale.sims}
\title{Scale simulated concentrations by the estimated emission rate of the nearest naive event}
\usage{
\method{scale}{sims}(times, sims, spikes, loc.est.all.events, rate.est.all.events)
}
\description{
Scale simulated concentrations by the estimated emission rate of the nearest naive event
}
