% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_durations.R
\name{estimate_durations}
\alias{estimate_durations}
\title{Estimate emission durations using CMS, accounting for CMS non-detect times}
\usage{
estimate_durations(data, directory, events, plot, time)
}
\arguments{
\item{data}{List. Forward model output data. The format must match that of the sample (default) shortened data from the 2023 ADED controlled release experiment, which is released with the \code{duration} package.}

\item{tz}{Character. Time zone. Default set to \code{America/Denver}.}

\item{directory}{Character. Location where results output is saved. If not specified, a directory is created and the location is shared with the user.}

\item{events}{Logical. Print the number of events per year? Default set to \code{FALSE}.}

\item{plot}{Logical. Return a plot of the distribution of average duration estimates by source? Default set to \code{FALSE}.}

\item{time}{Logical. Return the timing of the simulation. Default set to \code{TRUE}.}
}
\value{
Listed output of estimated durations by emission source. Saved in a newly created directory (\code{Results Output}).
}
\description{
Estimate emission durations using CMS, accounting for CMS non-detect times
}
\examples{
estimate_durations(directory = "Desktop", plot = TRUE)
}
\author{
William Daniels, Philip Waggoner, Dorit Hammerling
}
